backend/requirements.txt
django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pillow==10.4.0
nh3==0.2.21
pytest-django==4.9.0
